Hoist post fetching to parent
